requirements.txt
rich>=13.0.0
pydantic>=2.0.0
httpx>=0.25.0
python-dotenv>=1.0.0
click>=8.0.0
mcp>=0.1.0
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
readline>=6.2.4.1
pyyaml>=6.0
keyring>=24.0.0
